feat: modified icons and layout. feat: moved global logic into NavHolderActivity.
