+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.AppCompat.Light.NoActionBar">
+        android:theme="@style/AppTheme">
         <activity android:name=".activity.NavHolderActivity"></activity>
         <activity android:name=".activity.AlarmAndLocateResponseActivity" />
         <activity android:name=".activity.MainActivity">
