+++ app/src/main/res/drawable/round_people_alt_24.xml
@@ -1,9 +1,8 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
+    android:width="36dp"
+    android:height="36dp"
     android:viewportWidth="24"
-    android:viewportHeight="24"
-    android:tint="?attr/colorControlNormal">
+    android:viewportHeight="24">
     <path
         android:fillColor="@android:color/white"
         android:pathData="M16.67,13.13C18.04,14.06 19,15.32 19,17v3h3c0.55,0 1,-0.45 1,-1v-2C23,14.82 19.43,13.53 16.67,13.13z" />
