+++ app/src/main/res/drawable/settings_icon.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:name="vector"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+    <path
+        android:name="path"
+        android:pathData="M 19.14 12.94 C 19.18 12.64 19.2 12.33 19.2 12 C 19.2 11.68 19.18 11.36 19.13 11.06 L 21.16 9.48 C 21.34 9.34 21.39 9.07 21.28 8.87 L 19.36 5.55 C 19.24 5.33 18.99 5.26 18.77 5.33 L 16.38 6.29 C 15.88 5.91 15.35 5.59 14.76 5.35 L 14.4 2.81 C 14.36 2.57 14.16 2.4 13.92 2.4 L 10.08 2.4 C 9.84 2.4 9.65 2.57 9.61 2.81 L 9.25 5.35 C 8.66 5.59 8.12 5.92 7.63 6.29 L 5.24 5.33 C 5.02 5.25 4.77 5.33 4.65 5.55 L 2.74 8.87 C 2.62 9.08 2.66 9.34 2.86 9.48 L 4.89 11.06 C 4.84 11.36 4.8 11.69 4.8 12 C 4.8 12.31 4.82 12.64 4.87 12.94 L 2.84 14.52 C 2.66 14.66 2.61 14.93 2.72 15.13 L 4.64 18.45 C 4.76 18.67 5.01 18.74 5.23 18.67 L 7.62 17.71 C 8.12 18.09 8.65 18.41 9.24 18.65 L 9.6 21.19 C 9.65 21.43 9.84 21.6 10.08 21.6 L 13.92 21.6 C 14.16 21.6 14.36 21.43 14.39 21.19 L 14.75 18.65 C 15.34 18.41 15.88 18.09 16.37 17.71 L 18.76 18.67 C 18.98 18.75 19.23 18.67 19.35 18.45 L 21.27 15.13 C 21.39 14.91 21.34 14.66 21.15 14.52 L 19.14 12.94 Z M 12 15.6 C 10.02 15.6 8.4 13.98 8.4 12 C 8.4 10.02 10.02 8.4 12 8.4 C 13.98 8.4 15.6 10.02 15.6 12 C 15.6 13.98 13.98 15.6 12 15.6 Z"
+        android:fillColor="@android:color/black"
+        android:strokeWidth="1" />
+</vector>
