+++ app/src/main/res/layout/activity_nav_holder.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/root_layout"
+    android:id="@+id/home_root_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".activity.NavHolderActivity" />
