+++ app/src/main/res/menu/home_menu.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/home_menu_log"
+        android:title=""
+        android:icon="@drawable/round_schedule_24"
+        app:showAsAction="ifRoom" />
+    <item
+        android:id="@+id/home_menu_settings"
+        android:title="@string/settings_name"
+        android:icon="@drawable/settings_icon"
+        app:showAsAction="ifRoom" />
+</menu>
